backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
Werkzeug==2.3.7
python-dateutil==2.8.2
email-validator==2.1.0